Change how app args are passed to frame
